requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
bikram==2.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-filter==23.3
django-jazzmin==2.6.0
django-redis==5.4.0
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
flower==2.0.1
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
install==1.3.5
kombu==5.3.2
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
yapf==0.40.2
zipp==3.17.0
